Fix call to `rlp_memoverlap`
Because the address at `rlpEncodedOutput` is getting shifted after every RLP element is encoded, we need to check overlap of the *new* address plus the element size. Previously if a large element got added and the `rlpEncodedOutput` address shifted too much, this check would always fail because we were adding the total buffer size to the updated address.